Fixed template constexpr functions and added activation function and bias

+++ inc/connection.h
@@ -4,9 +4,9 @@ class Neuron;
 
 struct Connection
 {
-    Connection(const Neuron& neuron, double weight): inputNeuron(neuron), inputWeight(weight) {}
+    Connection(const Neuron& neuron, float weight): inputNeuron(neuron), inputWeight(weight) {}
     ~Connection() = default;
 
     const Neuron& inputNeuron;
-    double inputWeight;
+    float inputWeight;
 };
